Move canonicalPostURI into a separate utility file
