+++ src/utils/canonical-uri.js
@@ -1,0 +1,9 @@
+// Canonical post URI (pathname)
+export function canonicalURI(post) {
+  // If posted _only_ into groups, use first recipient's username
+  let urlName = post.createdBy.username;
+  if (post.recipients.length > 0 && !post.recipients.some((r) => r.type === 'user')) {
+    urlName = post.recipients[0].username;
+  }
+  return `/${encodeURIComponent(urlName)}/${encodeURIComponent(post.id)}`;
+}
